update db seeder and table parameter

+++ backend-api/database/migrations/2021_12_19_081725_create_schedules_table.php
@@ -27,8 +27,8 @@ class CreateSchedulesTable extends Migration
             $table->time('scheduleTime');
             $table->string('carModel');
             $table->string('carType');
-            $table->integer('timeEstimation');
-            $table->integer('priceEstimation');
+            $table->float('timeEstimation');
+            $table->float('priceEstimation');
             $table->string('scheduleStatus');
             $table->string('description')->nullable();
             $table->timestamps();
